+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessInstallPrePushHookTask.java
@@ -1,0 +1,46 @@
+package com.diffplug.gradle.spotless;
+
+import org.gradle.api.DefaultTask;
+import org.gradle.api.tasks.TaskAction;
+import org.gradle.work.DisableCachingByDefault;
+
+import com.diffplug.spotless.GitPrePushHookInstaller.GitPreHookLogger;
+import com.diffplug.spotless.GitPrePushHookInstallerGradle;
+
+/**
+ * A Gradle task responsible for installing a Git pre-push hook for the Spotless plugin.
+ * This hook ensures that Spotless formatting rules are automatically checked and applied
+ * before performing a Git push operation.
+ *
+ * <p>The task leverages {@link GitPrePushHookInstallerGradle} to implement the installation process.
+ */
+@DisableCachingByDefault(because = "not worth caching")
+public class SpotlessInstallPrePushHookTask extends DefaultTask {
+
+	/**
+	 * Executes the task to install the Git pre-push hook.
+	 *
+	 * <p>This method creates an instance of {@link GitPrePushHookInstallerGradle},
+	 * providing a logger to record informational and error messages during the installation process.
+	 * The installer then installs the hook in the root directory of the Gradle project.
+	 *
+	 * @throws Exception if an error occurs during the hook installation process.
+	 */
+	@TaskAction
+	public void performAction() throws Exception {
+		final var logger = new GitPreHookLogger() {
+			@Override
+			public void info(String format, Object... arguments) {
+				getLogger().lifecycle(String.format(format, arguments));
+			}
+
+			@Override
+			public void error(String format, Object... arguments) {
+				getLogger().error(String.format(format, arguments));
+			}
+		};
+
+		final var installer = new GitPrePushHookInstallerGradle(logger, getProject().getRootDir());
+		installer.install();
+	}
+}